Update: align test file with source change in this PR
